Fixing compile warning mostly

+++ src/main.cpp
@@ -9,10 +9,10 @@
 class GrapheneLatticeExtendedTest : public TwoDimensionalLattice {
  public:
   GrapheneLatticeExtendedTest()
-      : TwoDimensionalLattice(Vec<double>(3.0, 0), Vec<double>(0, sqrtf(3.0))) {
+      : TwoDimensionalLattice(Vec<double>(3.0, 0), Vec<double>(0, sqrt(3.0))) {
     AddSite(Site(Vec<double>{0, 0}));
-    AddSite(Site(Vec<double>{0.5, 0.5 * sqrtf(3.0)}));
-    AddSite(Site(Vec<double>{1.5, 0.5 * sqrtf(3.0)}));
+    AddSite(Site(Vec<double>{0.5, 0.5 * sqrt(3.0)}));
+    AddSite(Site(Vec<double>{1.5, 0.5 * sqrt(3.0)}));
     AddSite(Site(Vec<double>{2.0, 0}));
     AddEdge(Edge({0, 0}, 0, 1));
     AddEdge(Edge({0, 0}, 1, 2));
@@ -27,7 +27,7 @@ class GrapheneLatticeExtendedTest : public TwoDimensionalLattice {
 class TriangularLatticeTest : public TwoDimensionalLattice {
  public:
   TriangularLatticeTest(double a = 1.0)
-      : TwoDimensionalLattice(Vec<double>(a, 0), Vec<double>(0.5 * a, 0.5 * a * sqrtf(3.0))) {
+      : TwoDimensionalLattice(Vec<double>(a, 0), Vec<double>(0.5 * a, 0.5 * a * sqrt(3.0))) {
     AddSite(Site(Vec<double>{0, 0}));
     AddEdge(Edge({1, 0}, 0, 0));
     AddEdge(Edge({0, 1}, 0, 0));
@@ -38,10 +38,10 @@ class TriangularLatticeTest : public TwoDimensionalLattice {
 class GrapheneLattice : public TwoDimensionalLattice {
  public:
   GrapheneLattice()
-      : TwoDimensionalLattice(Vec<double>(1.5, 0.5 * sqrtf(3.0)),
-                              Vec<double>(1.5, -0.5 * sqrtf(3.0))) {
+      : TwoDimensionalLattice(Vec<double>(1.5, 0.5 * sqrt(3.0)),
+                              Vec<double>(1.5, -0.5 * sqrt(3.0))) {
     AddSite(Site(Vec<double>{0, 0}));
-    AddSite(Site(Vec<double>{0.5, 0.5 * sqrtf(3.0)}));
+    AddSite(Site(Vec<double>{0.5, 0.5 * sqrt(3.0)}));
     AddEdge(Edge({0, 0}, 0, 1));
     AddEdge(Edge({1, 0}, 1, 0));
     AddEdge(Edge({1, -1}, 1, 0));
